requirements.txt
# Annotation Tool v2 Dependencies
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
werkzeug==3.0.1
gunicorn==21.2.0

# Document processing
python-docx==1.1.0
PyMuPDF==1.23.8

# Logging
structlog==24.1.0

# Testing
pytest==8.0.0
pytest-cov==4.1.0

# Validation
jsonschema==4.21.1
